refactor: move user related api calls from profile-viewer-api-service to shared auth-lib-service
